% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcurrent.R
\name{getcurrent}
\alias{getcurrent}
\title{Get current value}
\usage{
getcurrent(ticker = "GOOG")
}
\arguments{
\item{ticker}{stock ticker symbol. E.g. "GOOG".}
}
\description{
Returns a row with the latest price for a given stock.
}
