% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googledata.R
\name{googledata}
\alias{googledata}
\title{Google data}
\usage{
googledata(ticker, from, to)
}
\arguments{
\item{ticker}{stock ticker symbol. E.g. "GOOG".}

\item{from}{start date. Either string or date object.}

\item{to}{end date. Either string or date object.}
}
\value{
dataframe with historical prices
}
\description{
Download historical prices for a given stock from Google
}
