% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iexdata.R
\name{iexdata}
\alias{iexdata}
\title{IEX data}
\usage{
iexdata(ticker, from = NULL, to = NULL)
}
\arguments{
\item{ticker}{stock ticker symbol. E.g. "GOOG".}

\item{from}{start date. Either string or date object.}

\item{to}{end date. Either string or date object.}
}
\value{
dataframe with historical prices
}
\description{
Download historical prices for a given stock from \url{https://iextrading.com}
}
