\name{listbyindustry}
\alias{listbyindustry}
\title{List stocks by industry}
\usage{
  listbyindustry()
}
\description{
  This function is specifically made for the ExtJS based
  OpenCPU app. It converts the 'stocktable' dataframe to a
  tree structure that can be loaded by the ExtJS tree
  store. It uses the 'stocktable' data object that ships
  with the package.
}

